backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
openai-whisper==20240930
transformers==4.46.0
sentence-transformers==3.2.1
torch==2.5.0
torchaudio==2.5.0
spacy==3.8.2
firebase-admin==6.6.0
reportlab==4.2.5
python-jose[cryptography]==3.3.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
pydantic==2.9.2
kaggle==1.6.14
